+++ environment.yml
@@ -4,8 +4,11 @@ channels:
 dependencies:
   - python >=3.10
   # Library Dependencies
+  - aiohttp
+  - dask
   - fsspec
   - numpy
+  - requests
   - xarray >=2024.10
   - zarr
   # Development Dependencies - Tools
