+++ pyproject.toml
@@ -21,8 +21,11 @@ keywords = [
 license = {text = "Apache-2.0"}
 requires-python = ">=3.10"
 dependencies = [
+  "aiohttp",
+  "dask",
   "fsspec",
   "numpy",
+  "requests",
   "xarray>=2024.10",
   "zarr"
 ]
