--- tests/__init__.py
@@ -1,3 +1,0 @@
-#  Copyright (c) 2025 by EOPF Sample Service team and contributors
-#  Permissions are hereby granted under the terms of the Apache 2.0 License:
-#  https://opensource.org/license/apache-2-0.
